# Optimize rails_xss_friendly_size and rails_friendly_size with ARM Neon instructions.