Note: This contains the bugfix from [this PR](https://github.com/ohler55/oj/pull/965).

Optimizing `Oj.dump(str, mode: :rails)`

The two benchmarks:
```
benchmark_encoding "mixed utf8", ([("a" * 5000) + "€" + ("a" * 5000)] * 500)
benchmark_encoding "ascii", ([("a" * 5000) + ("a" * 5000)] * 500)
```

The `€` triggers the `has_hi` so the speedup is quite different between the two benchmarks. The `mixed utf8` benchmark is a small speedup as the code spends most of its time in the main loop in `oj_dump_cstr`. 

## Scalar
```
== Encoding mixed utf8 (5003001 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [arm64-darwin24]
Warming up --------------------------------------
                  oj    12.000 i/100ms
Calculating -------------------------------------
                  oj    134.687 (± 2.2%) i/s    (7.42 ms/i) -    684.000 in   5.081051s

== Encoding ascii (5001501 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [arm64-darwin24]
Warming up --------------------------------------
                  oj    34.000 i/100ms
Calculating -------------------------------------
                  oj    330.226 (± 4.5%) i/s    (3.03 ms/i) -      1.666k in   5.055125s
```

## ARM Neon
```
== Encoding mixed utf8 (5003001 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [arm64-darwin24]
Warming up --------------------------------------
                  oj    15.000 i/100ms
Calculating -------------------------------------
                  oj    147.281 (± 5.4%) i/s    (6.79 ms/i) -    735.000 in   5.006387s

== Encoding ascii (5001501 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [arm64-darwin24]
Warming up --------------------------------------
                  oj    48.000 i/100ms
Calculating -------------------------------------
                  oj    469.450 (± 3.4%) i/s    (2.13 ms/i) -      2.352k in   5.016355s
```