@samyron 2025-05-03
I created another test file with longer strings which I can include in this if you'd like. With the ARM Neon optimized code strings need to be 16 bytes or longer otherwise that code will not be hit.

Additionally, I can also include an optimization of `rails_friendly_size` too. I can include both on this PR later tonight or I can create a follow up PR if you prefer.