@samyron 2025-05-04
This is ready for review. I added a bunch of tests which ensure the ARM Neon code gets executed. Testing `rails_friendly_size` is tricky though. I had to comment out code that handled that in order to get CI to build on ubuntu.